add target brightness model

+++ Strange_Lovers/DrivenPart/TargetLightValueControl/LightVal.h
@@ -6,7 +6,7 @@
 #include "../../EncoderPart/BrightnessEncoder.h"
 
 typedef struct{
-	int targLightVal;
+	float targLightVal;
 }LightVal;
 
 //LightValEncoder mLightValEncoder;
@@ -14,9 +14,9 @@ typedef struct{
 //LightValRange mLightValRange;
 
 extern void LV_init(LightVal *this_LightVal,BrightnessEncoder LVE);
-extern int LV_getLightVal(LightVal *this_LightVal);
-extern int LV_getTargLightVal(LightVal *this_LightVal);
-extern void LV_setTargLightVal(LightVal *this_LightVal,int parm);
+extern float LV_getLightVal(LightVal *this_LightVal);
+extern float LV_getTargLightVal(LightVal *this_LightVal);
+extern void LV_setTargLightVal(LightVal *this_LightVal,float parm);
 extern LightValRange LV_getLightValRange(LightVal *this_LightVal);
 extern void LV_setLightValListener(LightVal *this_LightVal,LightValListener LVL);
 extern void LV_setLightValRange(LightVal *this_LightVal,LightValRange LVR);
