+++ Strange_Lovers/StrategyPart/TargetValues.h
@@ -2,7 +2,7 @@
 #define STRATEGYPART_TARGET_VALUES_H
 
 typedef struct {
-	int target_brightness;
+	float target_brightness;
 	float target_curvature;
 	int target_speed;
 }TargetValues;
